Update integration test with binding and minor fixed
