+++ binding/stakeholdergroup.go
@@ -12,14 +12,14 @@ type StakeholderGroup struct {
 }
 
 //
-// Create an StakeholderGroup.
+// Create a StakeholderGroup.
 func (h *StakeholderGroup) Create(r *api.StakeholderGroup) (err error) {
 	err = h.client.Post(api.StakeholderGroupsRoot, &r)
 	return
 }
 
 //
-// Get an StakeholderGroup by ID.
+// Get a StakeholderGroup by ID.
 func (h *StakeholderGroup) Get(id uint) (r *api.StakeholderGroup, err error) {
 	r = &api.StakeholderGroup{}
 	path := Path(api.StakeholderGroupRoot).Inject(Params{api.ID: id})
@@ -36,7 +36,7 @@ func (h *StakeholderGroup) List() (list []api.StakeholderGroup, err error) {
 }
 
 //
-// Update an StakeholderGroup.
+// Update a StakeholderGroup.
 func (h *StakeholderGroup) Update(r *api.StakeholderGroup) (err error) {
 	path := Path(api.StakeholderGroupRoot).Inject(Params{api.ID: r.ID})
 	err = h.client.Put(path, r)
@@ -44,7 +44,7 @@ func (h *StakeholderGroup) Update(r *api.StakeholderGroup) (err error) {
 }
 
 //
-// Delete an StakeholderGroup.
+// Delete a StakeholderGroup.
 func (h *StakeholderGroup) Delete(id uint) (err error) {
 	err = h.client.Delete(Path(api.StakeholderGroupRoot).Inject(Params{api.ID: id}))
 	return
